added api routes for login and controller
