+++ client/src/utils/API.js
@@ -1,20 +1,13 @@
 import axios from "axios";
 
 export default {
-  // Gets all books
-  getBooks: function() {
-    return axios.get("/api/books");
+  // Get all users
+  getUser: function() {
+    return axios.get("/api/login");
   },
-  // Gets the book with the given id
-  getBook: function(id) {
-    return axios.get("/api/books/" + id);
-  },
-  // Deletes the book with the given id
-  deleteBook: function(id) {
-    return axios.delete("/api/books/" + id);
-  },
-  // Saves a book to the database
-  saveBook: function(bookData) {
-    return axios.post("/api/books", bookData);
+  
+  // Saves a User to the database
+  saveUser: function(userData) {
+    return axios.post("/api/signup", userData);
   }
 };
